Be more strict about the format of "disabled"
While we could be more tolerant here than in the spec, the risk is
that the majority of developers will not read the spec and will
instead try something with our verifier and expect that it will
always continue to work the same way.

In other words, if we're too lenient at the start, we will be
stuck supporting an alternative syntax forever and will impose that
on third-party verifiers as well.